+++ karma.conf.js
@@ -7,14 +7,14 @@ module.exports = function (config) {
       'tests/base_test_function.js',
       {pattern: 'tests/**/*.spec.js', included: true},
     ],
-    exclude: [
-      "tests/generate_data.spec.js"
-    ],
+    // exclude: [
+    //   "tests/generate_data.spec.js"
+    // ],
     reporters: ['progress'],
     port: 9876,  // karma web server port
     colors: true,
     logLevel: config.LOG_INFO,
-    browsers: [ 'Chrome', 'ChromeHeadless', 'Firefox'],
+    browsers: [ 'Chrome'],
     autoWatch: false,
     concurrency: Infinity,
     customLaunchers: {
