+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "di-web-analytics",
-  "version": "0.8.4",
+  "version": "0.8.5",
   "description": "Data insider web analytics",
   "repository": {
     "type": "git",
