+++ src/service/index.ts
@@ -4,3 +4,5 @@ export * from './analytics_core';
 export * from './di_analytics';
 export * from './logger';
 export * from './tracking_service';
+export * from './disable_analytics_core'
+export * from './analytics_core_impl'
